Added isCallable for use instead of isFunction
'isFunction x' tells us whether or not 'x' is a Nix function
value. Unfortunately, our functions can get messed around with, e.g. if they're
returned from a 'callPackage' invocation; in which case, they're replaced by an
attribute set containing 'override', 'overrideDerivation' and
'__functor' (hooray, another overloaded meaning for the word "functor"!). In
these cases, Nix *will* allow us to call this attribute set *as if it were* a
function, but unfortunately 'isFunction' will say it's not a function.

This is annoying if, for example, we want to use assertions to ensure that our
functions have been called correctly. In these cases, using 'isFunction' is too
strict, since we only really care if something is "callable" (i.e. can be
called *as if it were* a function), so these "functor" things should be
acceptable.

That's what this 'isCallable' function will check :)