Prevent infinite loops by using 'super' dependencies for stableHackageDb
The stableHackageDb stuff uses 'cabal', which is a Haskell package, and can
hence get into an infinite loop as it tries to resolve dependencies, etc. This
happened when overriding the version of tasty we use (to avoid a known bug).

To prevent this, we use cabal from a fixed nixpkgs; mostly since those fixed
versions don't include our config, and hence avoid our overrides and break the
cycle.