Add delay-failure option to haskellPkgWithDeps
We're running cabal in a derivation to figure out dependencies, then importing
the result into Nix. If cabal fails to solve the dependencies, then Nix's
evaluation will abort.

This can be desirable some times (bailing out as early as possible), but not
always. For example, on Hydra we only want evaluation to fail if our Nix code is
dodgy; if it's a package that's dodgy (e.g. if its dependencies aren't solvable
by cabal) then we'd like that just that package to fail at build time, rather
than derailing the evaluation of everything else.

Setting 'delay-failure' to true (default is false) will spot cabal failures, and
produce a valid package, but which will fail at build time with a helpful error
message.