Propagate Haskell deps out to pandocPkgs
When we import from a derivation, the result doesn't have an explicit dependency
on that derivation. This is useful, since it lets us do things like checking the
HEAD of a git repo, whilst still being able to use cached versions if there's
been no change (if the resulting derivation *did* depend on the repo-checking
derivation its revision string came from, we would have to rebuild due to a
cache miss).

For something like checking a git repo this is fine, since it's relatively
quick. For something like running cabal2nix on all of a package's dependencies
it would be nice to cache them. Especially since those runCabal2nix derivations
may be the only things which depend on our Hackage snapshot, and we *really*
don't want that to get garbage collected when we still want to import things
from it.

For this reason, the various "haskellPkg*.nix" scripts now pass along the
derivations that they've imported, which they call "gcRoots".

The "pandocPkgs" package looks like the only thing to be using this
infrastructure at the moment, but in any case it now adds these gcRoots to its
own dependencies using 'withDeps'. Since pandocPkgs itself is just a link farm,
this doesn't add much of a burden due to rebuilds.