Add haskellRelease
This is intended to be used in release.nix files for Haskell projects. It spits
out a derivation which, when built, produces a Nix file. That Nix file defines a
whole bunch of derivations of the given Haskell project, using different nixpkgs
Haskell package sets, and taking dependencies from Hackage.