Try using zlib-bindings instead of zip-archive
We want to test that we can supply the Haskell zlib package as an argument, when
needed (as well as the system's zlib package, when that's needed). We were doing
this by testing that zip-archive, which depends on Haskell's zlib, builds. It
turns out that its test suite doesn't work in Nix, since it doesn't expect its
timestamps to be set to 1970.

This commit switches that test out for zlib-bindings instead, which will
hopefully not be broken on Nix.