Fix and simplify SSH workaround in withNix
Nix 2.x doesn't allow nixbld users to connect to nix-daemon. We were working
around this by SSHing to "nixbuildtrampoline@localhost" and having that account
execute the required commands.

Unfortunately that's fraught with problems: in particular, the
nixbuildtrampoline user might not have permission to access paths that our Nix
expressions reference. This is the case for Nix's temporary build directories,
e.g.

    echo 'import <nixpkgs> {}' > test.nix
    ssh nixbuildtrampoline@localhost nix-build -E "(import $PWD/test.nix).hello"

This will die, since nixbuildtrampoline won't have permission to access
$PWD (/tmp/nix-build-whatever).

This commit adds the above as a test. It also fixes the problem by using SSH to
tunnel the nix-daemon's socket as the nixbuildtrampoline user, rather than
executing commands as that user.

This is simpler and more robust, and it also solves the problem of permissions,
etc. since the Nix commands are now running as the nixbld user; it's only their
communication with nix-daemon that's intercepted and user-switched.

This also has the benefit that we don't have to content with the awful argument
quoting issues of 'sshpass' anymore.