Simplify haskellRelease
Rather than taking filter functions to decide which nixpkgs and haskellPackages
combos to keep, we just take a list of each one we want.

Rather than making a derivation which builds a release.nix for subsequent
import, we just screw it and define the resulting set.

Lots of indirection removed (mostly thanks to not producing a derivation).

A bunch of assertions added.

Note that this will cause more building at eval time, due to now using
import-from-derivation on the result of runCabal2nix calls. runCabal2nix uses
Cabal and cabal2nix, which can take a while to build from scratch (the latter
has a huge pile of dependencies for some reason). That's what we were trying to
avoid, but sometimes simplicity is preferable.