withNix: Expose tunnel socket env var
Rather than setting NIX_REMOTE=daemon in the environment, then providing
wrappers which override that to the tunnel socket unix://..., we now just expose
that tunnel socket in the environment.

This means that we don't have to wrap nix-* commands, and that other builders
which use withNix now utilise the same workaround (for example cabal2nix).

Removing the wrapper scripts prevents us checking for the socket's existence and
spewing our big explanation if not found. This might cause problems which are
hard to debug, for those who don't know about our hack. To try and remove this
confusion as quickly as possible, we now always embed the warning/explanation in
the environment, so anyone debugging should come across it.