Ignore problematic 'haskelldb-connect-hdbc-catchio-transformers' pkg
This seems to cause really big filenames in our hackage tarball, tar contents
seem to be limited to 100 character paths unless a workaround is
enabled (e.g. "gnu format"). After trying a bunch of these formats, they all
seemed to trigger an error from cabal2nix, so removing the file seems to be the
only solution for now.