Put wrapped scripts in directories, rather than directly in Nix store
It's a bad idea to put Python scripts directly into the Nix store, since they
seem to scan for modules in their parent directory, which can take a while since
the Nix store is pretty huge. It's better to put them in a directory instead.

Since I tend to use 'wrap' quite a lot for Python scripts, it makes sense to put
all wrapped scripts in directories: it's very little overhead, and doesn't seem
to interfere with anything else.