Patch /usr/bin/env shebangs given in scripts
NixOS 19.03 defaults to using chroot sandboxes, which is nice but seems to break
a bunch of my derivations. One reason is that sandboxes can't access external
paths, which includes /usr/bin/env.

To avoid this, we now replace it with ${coreutils}/bin/env, but only in certain
circumstances:

 - This only affects scripts created with 'wrap' (and hence 'mkBin' and 'run')
 - This behaviour can be avoided by passing patchShebangs = false to wrap
 - Only script arguments are patched, not file arguments
 - Only the first line of the given script is patched, not any subsequent
   occurrences (e.g. if one script generates another)

This seems to solve the particular issues I've been having. I may revisit this
if/when I hit any more problems.

Note that the main reason I don't want to get involved in patching files is
because they may be symlinks, or wrappers for other files, etc. and that would
be messy to chase down while avoiding edge cases.