Add onlineCheck
If we're checking whether a package is up to date (e.g. by polling the project's
Web site for the latest version number), we should allow the check to
shortcircuit if this boolean is false.

If we're evaluating something and want to avoid such optional checks (e.g. if
we're offline), then we can toggle it using an env var.