Add checkedRacket
We already provide checkRacket to see if the current system should use an old
version of the racket package (newer versions are broken on i686), but we caused
some redundancy in downstream projects (e.g. warbo-utilities and nix-config). In
particular, our current policy w.r.t. racket is:

 - Default to using 'self.racket' from the system.
 - Allow the system config to override 'racket' to a non-broken version, if
   needed.
 - Provide 'checkRacket' to check if such an override is needed.

The problem is that "standalone" uses of Racket, e.g. in release.nix files,
would need their own checks and overrides (since they're isolated from any
system-wide overrides). This caused redundancy.

We now provide checkedRacket, which can be used as a drop-in replacement for the
'racket' package if there's any possibility that a switch might be needed.

checkedRacket will:

 - Invoke checkRacket to see if an override is needed
 - Use a known-good old package if so
 - Use the system's racket if not

This prevents having to copypaste the same code around.