Remove latestGit, which allows impure dependencies
It's OK when used with `stable = { rev = "..."; sha256 = "..."; }`, but that can
be converted to builtins.fetchGit pretty easily.

In fact, builtins.fetchGit can fetch the latest HEAD too, so this is basically
obsolete. The reason I'm killing it, is due to its reliance on __noChroot, which
can make things unbuildable when using a strict sandbox; whereas fetchGit does
not depend on sandboxing at all (it can be disabled by requiring pure evaluation
but that's not as widespread)