callCabal2nixWithPlan: take cabal and ghc from given haskellPackages
The generated plan is forced to use versions of bootstrap libraries like base
that match the ghc in use. This was previously hard-coded to use the version
from nixpkgs; now it will take the ghc attribute from the given haskellPackages
argument, although a separate ghc argument can be supplied if wanted. The same
goes for cabal-install.