fetchTreeFromGitHub: Remove Nixpkgs dependency, use builtin builders
Removing the dependency on Nixpkgs is useful, since we want to use this to FETCH
Nixpkgs from github in the first place! Builtin builders let us do this, though
it's a bit convoluted, since we can unpack the tarball using 'unpack-channel',
but that puts the result in a subdirectory; so we have to calculate an extra
SHA1 for that wrapper.

We can do that now that we've got gitTreeSingleton