fetchGitIPFS: Try using <nix/fetchurl.nix> instead of fetchTree
There's an issue with fetchTree, where it doesn't check whether a store path
with the given narHash already exists; so it ends up using the tarball cache
instead. This results in decompressing things over and over.

Instead, <nix/fetchurl.nix> seems to avoid that by being a fixed-output
derivation. My initial experiments with this have worked, so we'll see if it's a
nicer experience overall.